backend/requirements.txt
fastapi
SQLAlchemy
psycopg2-binary
geoalchemy2
redis
scikit-learn
tensorflow
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
pytest
httpx
pytest-asyncio
pytest-cov
pytest-mock
alembic
uvicorn
starlette
geopy
shapely
python-dotenv
requests
python-multipart
